==> notebooks/Rwanda/6_Do_mosaic_clipping.ipynb <==
Notebook to mosaic tiles and clip to country boundary.

### Mosaic post-processed tiles

! gdal_merge.py -o Results/Land_cover_prediction_postprocessed_Rwanda_2021_tiles_mosaic.tif -co COMPRESS=Deflate -ot Byte Results/Land_cover_prediction_postprocessed_Rwanda_2021_tile_*.tif

### clip mosaics to country boundary

import subprocess
bound_file="Data/Rwanda_Boundary.shp"

mosaic="Results/Land_cover_prediction_Rwanda_2021_tiles_mosaic.tif"
mosaic_clipped="Results/Land_cover_prediction_Rwanda_2021_tiles_mosaic_clipped.tif"
gdal_cmd=["gdalwarp", "-of", "GTiff", '-cutline',bound_file,'-crop_to_cutline',
          '-co','COMPRESS=DEFLATE',mosaic,mosaic_clipped]
p1=subprocess.run(gdal_cmd)

mosaic="Results/Land_cover_probability_Rwanda_2021_tiles_mosaic.tif"
mosaic_clipped="Results/Land_cover_probability_Rwanda_2021_tiles_mosaic_clipped.tif"
gdal_cmd=["gdalwarp", "-of", "GTiff", '-cutline',bound_file,'-crop_to_cutline',
          '-co','COMPRESS=DEFLATE',mosaic,mosaic_clipped]
p1=subprocess.run(gdal_cmd)

mosaic="Results/Land_cover_prediction_postprocessed_Rwanda_2021_tiles_mosaic.tif"
mosaic_clipped="Results/Land_cover_prediction_postprocessed_Rwanda_2021_tiles_mosaic_clipped.tif"
gdal_cmd=["gdalwarp", "-of", "GTiff", '-cutline',bound_file,'-crop_to_cutline',
          '-co','COMPRESS=DEFLATE',mosaic,mosaic_clipped]
p1=subprocess.run(gdal_cmd)
